code/requirements.txt
torch>=2.2.0
torchvision>=0.17.0
wandb>=0.25.0
pandas>=2.1.0
numpy>=1.26.0
pyyaml>=6.0.1
matplotlib>=3.8.0
torchmetrics[detection]>=1.4.0
pycocotools>=2.0.7
